MGNLIMG-83: added @Override's
git-svn-id: file:///home/dlipp/subversion-migration/svn-repo-copy/enterprise/imaging/trunk@48348 a15fbcf9-bee4-0310-bafe-d02d26ad3b5f

+++ magnolia-module-imaging/src/main/java/info/magnolia/imaging/operations/text/TextStyle.java
@@ -40,7 +40,7 @@ public class TextStyle {
     }
 
     public float calculateCharSpacing() {
-        return (float) ((double) getCharSpacing() / TextStyle.CHARSPACING_DIV_FACTOR);
+        return (float) (getCharSpacing() / TextStyle.CHARSPACING_DIV_FACTOR);
     }
 
     //    public Font deriveFont() {
